message_edit: Show edit if row is in the current list.
- We had an issue of disapperaing editbox from the narrow
on rerender. issue pointed out here
https://chat.zulip.org/#narrow/stream/9-issues/topic/message.20editing.20bug/near/1163844

- This happens when a row with same `id`, but not
belonging to the current message_list (for ex in narrows:
we have same message id in home as well as the current
message_list but the rows are different) is getting appended
with the editbox of a message in the current list. The editbox
disappears and gets attached to the other row.

- This is fixed by checking whether we are trying to show edit
to a row with the id in the current message list and not in any
message list.