messagbox_focus: Fix message_box focus on page rerender.
- The focus from a messagebox was going away
because we are rerendering the page on update_message
event.

To fix this we do the following:
1. We track currently_focused_message_id in message_edit.js :
- Everytime an user starts editing a message we set it to
currently_focused_message_id
- we attach an on focus event listener to each message box. The
listener sets the currently_focused_message_id according to
the id to the message edit box clicked on.
- If there is an on click on something other than the message_box
we set currently_focused_message_id as undefined
- When the edit message box is removed from the dom, we check
if it had the currently_focused_message_id, in that case we
set currently_focused_message_id to undefined.

2. We update the focus to the currently_focused_message_id,
after rerender:
- We get currently_focused_message_id from message_edit.js
and we update the focus to the message_id after rerender.

Fixes #18130